news_app/requirements.txt
beautifulsoup4==4.12.2
celery==5.2.7
gunicorn==20.1.0
Django==4.2
django-celery-beat==2.5.0
googletrans==3.1.0a0
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
redis==4.5.4
requests==2.28.2
djangorestframework==3.14.0
drf-yasg==1.21.5
django-filter==23.2